+++ sdk/identity/azure-identity/azure/identity/aio/_credentials/managed_identity.py
@@ -7,7 +7,7 @@ import os
 from typing import TYPE_CHECKING
 
 from azure.core.credentials import AccessToken
-from azure.core.exceptions import HttpResponseError
+from azure.core.exceptions import ClientAuthenticationError, HttpResponseError
 from azure.core.pipeline.policies import AsyncRetryPolicy
 
 from azure.identity._credentials.managed_identity import _ManagedIdentityBase
@@ -129,7 +129,24 @@ class ImdsCredential(_AsyncManagedIdentityBase):
             params = {"api-version": "2018-02-01", "resource": resource}
             if self._client_id:
                 params["client_id"] = self._client_id
-            token = await self._client.request_token(scopes, method="GET", params=params)
+
+            try:
+                token = await self._client.request_token(scopes, method="GET", params=params)
+            except HttpResponseError as ex:
+                # 400 in response to a token request indicates managed identity is disabled,
+                # or the identity with the specified client_id is not available
+                if ex.status_code == 400:
+                    self._endpoint_available = False
+                    message = "ManagedIdentityCredential authentication unavailable. "
+                    if self._client_id:
+                        message += "The requested identity has not been assigned to this resource."
+                    else:
+                        message += "No identity has been assigned to this resource."
+                    raise CredentialUnavailableError(message=message) from ex
+
+                # any other error is unexpected
+                raise ClientAuthenticationError(message=ex.message, response=ex.response) from None
+
         return token